split out cache
